Added Ownable interface, Color field for Players, began SVG generator.
An Ownable interface was created, which is used for all
BoardCells which can be owned.
Players now have a color field, which will be used later to
set the background color of owned cells.
Began work on the SVG generator, which will make images of
properties, depending on who owns the property and other things.
Also re-built javadocs.